mitosheet: cleanup and add documentation
