+++ mitosheet/src/components/toolbar/ToolbarDropdownSelector.tsx
@@ -3,15 +3,21 @@ import { ToolbarDropdowns, UIState } from '../../types';
 import { classNames } from '../../utils/classNames';
 
 
-interface ToolbarMenuProps {
+interface ToolbarDropdownSelectorProps {
     type: ToolbarDropdowns,
     uiState: UIState,
     setUIState: React.Dispatch<React.SetStateAction<UIState>>,
     children: JSX.Element
 }
 
-// TODO: this needs a better name
-const ToolbarMenu = (props: ToolbarMenuProps): JSX.Element => {
+/**
+ * A wrapper component around the selectors that create dropdowns in the toolbar
+ * that allows the user to see all the editing options Mito supports.
+ * 
+ * If one of these dropdowns is open and you mouse over a different dropdown, it will
+ * switch to that new one.
+ */
+const ToolbarDropdownSelector = (props: ToolbarDropdownSelectorProps): JSX.Element => {
 
     const selected = props.uiState.currOpenToolbarDropdown === props.type;
 
@@ -40,4 +46,4 @@ const ToolbarMenu = (props: ToolbarMenuProps): JSX.Element => {
     );
 }
 
-export default ToolbarMenu;
+export default ToolbarDropdownSelector;
